+++ spec/saml/kit/builders/authentication_request_spec.rb
@@ -25,5 +25,13 @@ RSpec.describe Saml::Kit::Builders::AuthenticationRequest do
       expect(result['AuthnRequest']['Issuer']).to eql(issuer)
       expect(result['AuthnRequest']['NameIDPolicy']['Format']).to eql(Saml::Kit::Namespaces::PERSISTENT)
     end
+
+    context "when force authn is enabled" do
+      before { subject.force_authn = true }
+
+      let(:result) { Hash.from_xml(subject.to_xml) }
+
+      specify { expect(result['AuthnRequest']['ForceAuthn']).to eql('true') }
+    end
   end
 end