+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 Version 1.0.31
+
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -6,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for ForceAuthn attribute on AuthnRequest
+
 ### Changed
 - Rescue from invalid signature validation
 - Change minimum ruby version to 2.4